data-service/requirements.txt
flask==2.0.1
werkzeug==2.0.3
kafka-python==2.0.2
psycopg2-binary==2.9.3
pytest==7.4.0